Capstone-Project-2/requirements.txt
boto3==1.34.146
pandas==2.2.2
python-dotenv==1.0.1
